# api/openapi.yaml
openapi: 3.0.3
info:
  title: CodeForge API
  description: Remote AI code task runner. Submit code tasks, monitor progress via SSE, and create pull requests.
  version: 1.0.0
  contact:
    name: CodeForge
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /api/v1/tasks/{taskID}/create-pr:
    post:
      summary: Create a pull request from task changes
      operationId: createPR
      tags: [Tasks]
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePRRequest"
      responses:
        "200":
          description: PR created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePRResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/v1/tasks/{taskID}/post-review:
    post:
      summary: Post review comments to PR/MR
      operationId: postReviewComments
      tags: [Tasks]
      description: |
        Posts the task's ReviewResult as comments to the associated GitHub PR or GitLab MR.
        Task must have a review_result (run a review first).
        Token is resolved from the task's provider_key or access_token.
      parameters:
        - name: taskID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pr_number:
                  type: integer
                  description: Override PR/MR number (defaults to task config or task field)
      responses:
        "200":
          description: Review comments posted
          content:
            application/json:
              schema:
                type: object
                properties:
                  review_url:
                    type: string
                    description: URL to the posted review on GitHub/GitLab
                  comments_posted:
                    type: integer
                    description: Number of line-level comments posted
                  pr_number:
                    type: integer
                    description: The PR/MR number used
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/task-types:
    get:
      summary: List available task types
      operationId: listTaskTypes
      tags: [Tasks]
      description: |
        Returns the list of available task types with their descriptions.
        Use this to populate task type selectors in the UI.
      responses:
        "200":
          description: List of task types
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_types:
                    type: array
                    items:
                      $ref: "#/components/schemas/TaskTypeInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Operation conflicts with current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
...
    TaskTypeInfo:
      type: object
      properties:
        name:
          type: string
          description: Task type identifier
          example: plan
        label:
          type: string
          description: Human-readable label
          example: Plan
        description:
          type: string
          description: What this task type does
          example: "Analyze the codebase and create an implementation plan without modifying files"
...
    CreatePRRequest:
      type: object
      properties:
        title:
          type: string
          description: PR title (auto-generated if empty)
        description:
          type: string
          description: PR description (auto-generated if empty)
        target_branch:
          type: string
          description: Target branch (defaults to main/master)
    CreatePRResponse:
      type: object
      properties:
        pr_url:
          type: string
        pr_number:
          type: integer
        branch:
          type: string
...